ctld: Move kernel_limits into login.c where it is used
Reviewed by:	asomers
Sponsored by:	Chelsio Communications
Differential Revision:	https://reviews.freebsd.org/D48771
