+++ usr.sbin/ctld/kernel.c
@@ -77,7 +77,7 @@
 
 extern bool proxy_mode;
 
-static int	ctl_fd = 0;
+int	ctl_fd = 0;
 
 void
 kernel_init(void)
@@ -904,60 +904,6 @@ kernel_handoff(struct ctld_connection *conn)
 	if (req.status != CTL_ISCSI_OK) {
 		log_errx(1, "error returned from CTL iSCSI handoff request: "
 		    "%s; dropping connection", req.error_str);
-	}
-}
-
-void
-kernel_limits(const char *offload, int s, int *max_recv_dsl, int *max_send_dsl,
-    int *max_burst_length, int *first_burst_length)
-{
-	struct ctl_iscsi req;
-	struct ctl_iscsi_limits_params *cilp;
-
-	bzero(&req, sizeof(req));
-
-	req.type = CTL_ISCSI_LIMITS;
-	cilp = (struct ctl_iscsi_limits_params *)&(req.data.limits);
-	if (offload != NULL) {
-		strlcpy(cilp->offload, offload, sizeof(cilp->offload));
-	}
-	cilp->socket = s;
-
-	if (ioctl(ctl_fd, CTL_ISCSI, &req) == -1) {
-		log_err(1, "error issuing CTL_ISCSI ioctl; "
-		    "dropping connection");
-	}
-
-	if (req.status != CTL_ISCSI_OK) {
-		log_errx(1, "error returned from CTL iSCSI limits request: "
-		    "%s; dropping connection", req.error_str);
-	}
-
-	if (cilp->max_recv_data_segment_length != 0) {
-		*max_recv_dsl = cilp->max_recv_data_segment_length;
-		*max_send_dsl = cilp->max_recv_data_segment_length;
-	}
-	if (cilp->max_send_data_segment_length != 0)
-		*max_send_dsl = cilp->max_send_data_segment_length;
-	if (cilp->max_burst_length != 0)
-		*max_burst_length = cilp->max_burst_length;
-	if (cilp->first_burst_length != 0)
-		*first_burst_length = cilp->first_burst_length;
-	if (*max_burst_length < *first_burst_length)
-		*first_burst_length = *max_burst_length;
-
-	if (offload != NULL) {
-		log_debugx("Kernel limits for offload \"%s\" are "
-		    "MaxRecvDataSegment=%d, max_send_dsl=%d, "
-		    "MaxBurstLength=%d, FirstBurstLength=%d",
-		    offload, *max_recv_dsl, *max_send_dsl, *max_burst_length,
-		    *first_burst_length);
-	} else {
-		log_debugx("Kernel limits are "
-		    "MaxRecvDataSegment=%d, max_send_dsl=%d, "
-		    "MaxBurstLength=%d, FirstBurstLength=%d",
-		    *max_recv_dsl, *max_send_dsl, *max_burst_length,
-		    *first_burst_length);
 	}
 }
